requirement.txt
## The following requirements were added by pip freeze:
asgiref==3.5.1
attrs==21.4.0   
autobahn==22.4.2
Automat==20.2.0
cffi==1.15.0
channels-redis
channels==3.0.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
zope.interface==5.4.0